Divide into Prover, Verifier, and data classes. update docs. isolate PM algorithms for eventual removal.
