+++ src/main/kotlin/org/cryptobiotic/mixnet/core/VectorP.kt
@@ -50,68 +50,15 @@ fun Prod(vp: VectorP): ElementModP {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-// parallel Prod(powP)
-class PMProdPowP(val vp: VectorP, val exp: VectorQ, val nthreads: Int = 10) {
-    val manager = SubArrayManager(vp.nelems, nthreads)
-    init {
-        require (vp.nelems == exp.nelems)
-    }
-    var result = exp.group.ONE_MOD_P
-
-    fun calc(): ElementModP {
-        runBlocking {
-            val jobs = mutableListOf<Job>()
-            val workProducer = producer(manager)
-            repeat(nthreads) {
-                jobs.add( launchCalculator(workProducer) {  subarray -> powp(subarray) } )
-            }
-            // wait for all calculations to be done, then close everything
-            joinAll(*jobs.toTypedArray())
-        }
-        return result
-    }
-
-    private fun CoroutineScope.producer(manager : SubArrayManager): ReceiveChannel<Int> =
-        produce {
-            repeat(manager.nthreads) { subidx ->
-                if ( manager.size[subidx] > 0) {
-                    send(subidx)
-                    yield()
-                }
-            }
-            channel.close()
-        }
-
-    private val mutex = Mutex()
-
-    private fun CoroutineScope.launchCalculator(
-        input: ReceiveChannel<Int>,
-        calculate: (Int) -> ElementModP
-    ) = launch(Dispatchers.Default) {
-
-        for (subidx in input) {
-            val calcResult = calculate(subidx)
-            mutex.withLock {
-                result *= calcResult
-            }
-            yield()
-        }
-    }
-
-    fun powp(subidx: Int): ElementModP {
-        var result = exp.group.ONE_MOD_P
-        for (rowidx in manager.subarray(subidx)) {
-            result *= vp.elems[rowidx] powP exp.elems[rowidx]
-        }
-        return result
-    }
+fun prodPowP(generators: VectorP, exps: VectorQ, nthreads: Int = 10): ElementModP {
+    return if (nthreads == 0) Prod(generators powP exps)           // CE n exp, 1 acc
+    else PProdPowP(generators, exps, nthreads).calc()
 }
 
 class PProdPowP(val vp: VectorP, val exp: VectorQ, val nthreads: Int = 10) {
     var result = exp.group.ONE_MOD_P
 
     fun calc(): ElementModP {
-
         runBlocking {
             val jobs = mutableListOf<Job>()
             val pairProducer = producer(vp, exp)
@@ -121,7 +68,6 @@ class PProdPowP(val vp: VectorP, val exp: VectorQ, val nthreads: Int = 10) {
             // wait for all calculations to be done, then close everything
             joinAll(*jobs.toTypedArray())
         }
-
         return result
     }
